[-] MO : themeconfigurator : demo images are no more deleted when uninstalling the module.

+++ modules/themeconfigurator/themeconfigurator.php
@@ -37,7 +37,7 @@ class ThemeConfigurator extends Module
 	{
 		$this->name = 'themeconfigurator';
 		$this->tab = 'front_office_features';
-		$this->version = '0.2';
+		$this->version = '0.3';
 		$this->bootstrap = true;
 		$this->secure_key = Tools::encrypt($this->name);
 		$this->default_language = Language::getLanguage(Configuration::get('PS_LANG_DEFAULT'));
@@ -335,7 +335,7 @@ class ThemeConfigurator extends Module
 		if (realpath(dirname($file_name)) != realpath($this->uploads_path))
 			Tools::dieOrLog(sprintf('Could not find upload directory'));
 
-		if ($image != '' && is_file($file_name))
+		if ($image != '' && is_file($file_name) && !strpos($file_name, 'banner-img') && !strpos($file_name, 'bg-theme') && !strpos($file_name, 'footer-bg'))
 			unlink($file_name);
 	}
